+++ crates/pathbase-client/src/lib.rs
@@ -1,0 +1,3 @@
+#![doc = include_str!("../README.md")]
+
+include!(concat!(env!("OUT_DIR"), "/pathbase_client.rs"));